+++ CloudShell.ResourceDefinitions.ReferenceProviders/ExecutableApplicationConfiguration.cs
@@ -1,0 +1,6 @@
+namespace CloudShell.ResourceDefinitions.ReferenceProviders;
+
+public sealed record ExecutableApplicationConfiguration(
+    string Path,
+    string? Arguments,
+    string? WorkingDirectory = null);